AutomationFolder/requirements.txt
matplotlib
pandas
cupy-cuda12x
openpyxl
tqdm
pillow
opencv-python
XlsxWriter